chore: modify recruit page
