+++ packages/gdsc-dju-web/src/apis/pageData/recruitDetails.ts
@@ -1,4 +1,36 @@
 import { RecruitmentDetailType } from '../../types/pageData';
+export const recruitDefaultText = {
+  developer:
+    '코딩에 대한 남다른 애정을 가지신 분\n' +
+    '1년이상 활동이 가능하신 분\n' +
+    '책임감이 있으신 분\n' +
+    '주도적으로 의견을 얘기하거나 반박하는데 거리낌이 없으신 분\n' +
+    '동료들에게 좋은 시너지를 줄 수 있으신 분\n' +
+    '타인의 의견을 적극적으로 수용하고 개선하려고 노력하시는 분\n' +
+    '프로젝트의 애정과 책임을 가지고 주도적으로 개발, 개선할 수 있으신 분\n' +
+    '커뮤니케이션 및 개발 문서 정리에 능숙하신 분\n' +
+    '코딩에 대하여 잘 모르더라도 매일 학습하고, 모르는 게 있으면 주도적으로 해결할 수 있으신 분',
+  designer:
+    'UX/UI 디자인에 대한 남다른 애정을 가지신 분\n' +
+    '1년이상 활동이 가능하신 분\n' +
+    '책임감이 있으신 분\n' +
+    '주도적으로 의견을 얘기하거나 반박하는데 거리낌이 없으신 분\n' +
+    '동료들에게 좋은 시너지를 줄 수 있으신 분\n' +
+    '타인의 의견을 적극적으로 수용하고 개선하려고 노력하시는 분\n' +
+    '프로젝트의 애정과 책임을 가지고 주도적으로 개발, 개선할 수 있으신 분\n' +
+    '커뮤니케이션 및 문서 정리에 능숙하신 분\n' +
+    '디자인에 대하여 잘 모르더라도 매일 학습하고, 모르는 게 있으면 주도적으로 해결할 수 있으신 분',
+  goodMember:
+    'Github, git에 대한 경험이나 프로젝트 경험이 있는 분이면 좋습니다.\n' +
+    'Framework나 Library 사용이 능숙하신 분이면 좋습니다.',
+  benefits:
+    '여러 국가의 GDSC 및 GDSC Korea와의 커뮤니케이션\n' +
+    'GDG 개발자와의 컨택 기회\n' +
+    'google에서 지원하는 기념품\n' +
+    '자체 스터디를 위한 google의 코세라, 퀵랩 바우처 지급',
+  process:
+    '서류접수 > 인터뷰 > 최종합격 (결과 발표에는 최대 2주가 소요될 수 있습니다.)',
+};
 
 export const RecruitDetails: RecruitmentDetailType[] = [
   {
@@ -25,7 +57,7 @@ export const RecruitDetails: RecruitmentDetailType[] = [
     introduction:
       'Beginner 포지션이 개설된 이후 포지션의 기준이 변경되었어요.\n' +
       '프레임워크를 바탕으로 하나의 프로젝트를 혼자 개발하실 수 있는 분이 포지션을 얻을 수 있어요.\n' +
-      '동료 개발자와 협업하거나 열정적인 분들과 함께 성장해나가고 싶은 분들은 환영해요.',
+      '동료 개발자와 협업하거나 열정적인 분들과 함께 성장해나가고 싶은 분들을 환영해요.',
     activity:
       '2주마다 있는 정기세션에 참여할 수 있어요.\n' +
       'GDSC KOR나 GDSC DJU가 기획 및 운영하는 이벤트에 운영진으로 참여할 수 있어요.\n' +
@@ -61,7 +93,7 @@ export const RecruitDetails: RecruitmentDetailType[] = [
     introduction:
       'Beginner 포지션이 개설된 이후 포지션의 기준이 변경되었어요.\n' +
       '프레임워크를 바탕으로 하나의 프로젝트를 혼자 개발하실 수 있는 분이 포지션을 얻을 수 있어요.\n' +
-      '동료 개발자와 협업하거나 열정적인 분들과 함께 성장해나가고 싶은 분들은 환영해요.',
+      '동료 개발자와 협업하거나 열정적인 분들과 함께 성장해나가고 싶은 분들을 환영해요.',
   },
   {
     id: 'backend',
@@ -90,7 +122,7 @@ export const RecruitDetails: RecruitmentDetailType[] = [
     introduction:
       'Beginner 포지션이 개설된 이후 포지션의 기준이 변경되었어요.\n' +
       '프레임워크를 바탕으로 하나의 프로젝트를 혼자 개발하실 수 있는 분이 포지션을 얻을 수 있어요.\n' +
-      '동료 개발자와 협업하거나 열정적인 분들과 함께 성장해나가고 싶은 분들은 환영해요.',
+      '동료 개발자와 협업하거나 열정적인 분들과 함께 성장해나가고 싶은 분들을 환영해요.',
     activity:
       '2주마다 있는 정기세션에 참여할 수 있어요.\n' +
       'GDSC KOR나 GDSC DJU가 기획 및 운영하는 이벤트에 운영진으로 참여할 수 있어요.\n' +
@@ -121,7 +153,7 @@ export const RecruitDetails: RecruitmentDetailType[] = [
       '최소 하나 이상의 ML 라이브러리(sklearn, TensorFlow, OpenCV 등)를 익숙하게 활용하실 수 있으신 분 ',
     introduction:
       '진로가 머신러닝 엔지니어이신 분들을 위해 개설되었어요.\n' +
-      '동료 개발자와 협업하거나 프로젝트를 진행하는 경험을 얻고 싶은 분들은 환영해요.',
+      '동료 개발자와 협업하거나 프로젝트를 진행하는 경험을 얻고 싶은 분들을 환영해요.',
     activity:
       '2주마다 있는 정기세션에 참여할 수 있어요.\n' +
       'GDSC KOR나 GDSC DJU가 기획 및 운영하는 이벤트에 운영진으로 참여할 수 있어요.\n' +
@@ -150,7 +182,7 @@ export const RecruitDetails: RecruitmentDetailType[] = [
       'MarkUp(HTML)언어를 사용할 줄 아시는 분 ',
     introduction:
       '진로가 UX/UI 디자이너이신 분들을 위해 개설되었어요.\n' +
-      '개발자와 협업하거나 동료 디자이너와 협업하고 싶은 분들은 환영해요.',
+      '개발자와 협업하거나 동료 디자이너와 협업하고 싶은 분들을 환영해요.',
     activity:
       '2주마다 있는 정기세션에 참여할 수 있어요.\n' +
       'GDSC KOR나 GDSC DJU가 기획 및 운영하는 이벤트에 운영진으로 참여할 수 있어요.\n' +
